+++ PyIRI/tests/__init__.py
@@ -1,0 +1,20 @@
+"""Unit and Integration Tests for PyIRI.
+
+Note
+----
+This file must remain empty of code for pytest to function.
+
+Example
+-------
+To run all tests:
+::
+
+  pytest
+
+To run a specific test, include the filename (with the full path if not in the
+test directory), class, and test name:
+::
+
+  pytest test_igrf_lib.py::TestIGRFLibUtil::test_good_inc_verification
+
+"""